spawn classification json update
Run this with the learning json file(s) in the same folder, it'll fix the imcompletely classified spawn points (type 1) (see commit before this one) Also it'll classify the new type 1x60h3 correctly (pretty rare to have, also type 1 before, but can be distinguished from the incompletely classified ones). There is also a fix for a much earlier thing, which could lead to 1x60 spawns not being properly displayed on the map.

Last but not least it'll output a file containing all the information of the json files in the folder combined with duplicates removed. And it'll output the stats for it. This file may come in handy for a future update, if you have the data for an area you'll be able to freely select the center and radius in meters for the scan and it'll select the appropriate points from the combined file.

To do:
Actually integrate the new 1x60h3 spawn point type into the code.