Tool for visualizing complex scan areas
Currently you can create a polygon by clicking to add points. Control+click will close the polygon and tile it with hexagons, then save the results to a JSON file to use with the scanner in RAW mode.

There is also a function addHexGrid that can create a spiral around a center hexagon (the original scan method of PGO-mapscan). However, the interface doesn't implement a way to place these.